main/requirements.txt
streamlit==1.31.0
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.2.0
plotly==5.18.0
opencv-python-headless==4.9.0.80
scipy==1.11.4
requests==2.31.0